// status-page/src/types/enums.ts
export enum StatusPageKind {
  PUBLIC = "PUBLIC",
  CUSTOMER = "CUSTOMER",
  INTERNAL = "INTERNAL",
}

export enum ComponentStatus {
  OPERATIONAL = "OPERATIONAL",
  DEGRADED_PERFORMANCE = "DEGRADED_PERFORMANCE",
  PARTIAL_OUTAGE = "PARTIAL_OUTAGE",
  FULL_OUTAGE = "FULL_OUTAGE",
}

export enum StatusPageIncidentStatus {
  INVESTIGATING = "INVESTIGATING",
  IDENTIFIED = "IDENTIFIED",
  MONITORING = "MONITORING",
  RESOLVED = "RESOLVED",
}